+++ src/workers/sync/ItemState/domain/ItemDelta.ts
